# docs/northbound-openapi.yaml
openapi: 3.1.0
info:
  title: ClawManager Northbound API
  version: 1.2.0
  description: Public API for authenticated Lite instance provisioning and password-protected ShareLink management.
servers:
  - url: https://northbound.example.com:38443
paths:
...
  /api/northbound/v1/auth/refresh:
    post:
      summary: Rotate the refresh token and issue a new access token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
                  description: Current opaque refresh token. A successful refresh rotates it; clients must immediately replace the old value.
      responses:
        "200":
          description: Token rotated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          $ref: "#/components/responses/AuthInvalid"
...
  /api/northbound/v1/operations/{id}:
    get:
      summary: Get an asynchronous operation owned by the current user
      operationId: getOperation
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Opaque operation ID returned by createLiteInstance.
          schema: { type: string, minLength: 1 }
      responses:
        "200":
          description: Operation state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operation"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    TokenResponse:
      type: object
      required: [access_token, refresh_token, token_type, expires_in, refresh_expires_in, scopes, session_id]
      properties:
        access_token: { type: string, description: Opaque Bearer token for northbound API calls. }
        refresh_token: { type: string, description: Opaque rotating refresh token; replace the previous token after every successful refresh. }
        token_type: { type: string, const: Bearer }
        expires_in: { type: integer, minimum: 1, description: Access-token lifetime in seconds; default deployment value is 1800. }
        refresh_expires_in: { type: integer, minimum: 1, description: Refresh-token lifetime in seconds; default deployment value is 604800. }
        scopes:
          type: array
          description: Permissions fixed when the session is created; re-login is required to receive newly introduced scopes.
          items:
            type: string
            enum: [lite-instances:create, lite-instances:read, lite-instances:share-link:manage, lite-instances:share-link:reset]
        session_id: { type: string, description: Northbound session identifier for audit and logout; it is not an access token. }
...
    Operation:
      type: object
      required: [operation_id, status, resource_type, created_at, updated_at]
      properties:
        operation_id: { type: string, description: Opaque ID used to poll this operation. }
        status: { type: string, enum: [queued, processing, succeeded, failed] }
        resource_type: { type: string, const: lite_instance }
        instance_id: { type: integer, minimum: 1, description: Present after success; use it for instance and ShareLink endpoints. }
        error_code: { type: string }
        error_message: { type: string }
        created_at: { type: string, format: date-time }
        started_at: { type: [string, "null"], format: date-time }
        finished_at: { type: [string, "null"], format: date-time }
        updated_at: { type: string, format: date-time }
...
    Error:
      type: object
      required: [code, message, request_id]
      properties:
        code: { type: string }
        message: { type: string }
        request_id: { type: string }
  responses:
...
    AuthInvalid:
      description: Invalid, expired, or revoked northbound session
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    NotFound:
      description: Resource does not exist or is not owned by this user
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }